requirements.txt
fastapi
uvicorn
python-socketio
jinja2
aiohttp